// admin/src/api/generated/schemas/adminCurrencyFeatureSettings.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCurrencyFeatureSettings = {
  web2_deposit_enabled?: boolean;
  web2_withdraw_enabled?: boolean;
  web2_transfer_enabled?: boolean;
  web3_deposit_enabled?: boolean;
  web3_withdraw_enabled?: boolean;
  web3_swap_enabled?: boolean;
  use_global_withdraw_fee?: boolean;
  use_global_withdraw_audit?: boolean;
  use_global_transfer_audit?: boolean;
};