// admin/src/api/generated/schemas/adminCurrencyWithdrawAuditRuleItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCurrencyWithdrawAuditRuleItem = {
  id?: string;
  min_amount?: string;
  max_amount?: string;
  strategy?: string;
  enabled?: boolean;
  sort_order?: number;
};