// admin/src/api/generated/schemas/adminCurrencyWithdrawFeeRuleItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCurrencyWithdrawFeeRuleItem = {
  id?: string;
  rule_type?: string;
  value?: string;
  min_fee?: string;
  max_fee?: string;
  min_amount?: string;
  max_amount?: string;
  enabled?: boolean;
  sort_order?: number;
};